// generated/src/app/api-models/cap-pool.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapPool {
  capModelId: string;
  capEntityCode: string;
  capEntityId: string;
  seqCapPoolId: number;
  description: string;
  mbrThresholdFrom: number;
  mbrThresholdThru: number;
  descriptionPmpm: string;
  capRateId: string;
  pctCapRate: number;
  companyCodePmpm: string;
  glRefCodePmpm: string;
  pctWithholdPmpm: number;
  descriptionWhld: string;
  fileTypeWhld: string;
  companyCodeWhld: string;
  glRefCodeWhld: string;
  descriptionDed: string;
  fileTypeDed: string;
  companyCodeDed: string;
  glRefCodeDed: string;
  amountTypeDed: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  postingRulePmpm: string;
  postingRuleWhld: string;
  postingRuleDed: string;
  effectiveDate: Date;
  terminationDate: Date;
  capPoolId: string;
}